requirements.txt
open3d
numpy-stl
scipy
matplotlib
trimesh
